+++ apps/portfolio/composables/useBlog.ts
@@ -1,6 +1,5 @@
-import { useStoryblokApi } from '@storyblok/vue'
+import { useLogger } from '@alvarosabu/use'
 import { Story, StoryAsset, StoryContent, StoryStatus, StoryVersion } from './useStories'
-
 export interface ArticleStoryContent extends StoryContent {
   title: string
   media: StoryAsset
@@ -35,35 +34,35 @@ function formatArticleStory(story: ArticleStory): ArticleStory {
 }
 
 export function useBlog() {
-  const storyapi = useStoryblokApi()
-  const { logError } = useLogger()
+  /* const storyapi = useStoryblokApi() */
+  const { error } = useLogger('[ AS 🐧]')
 
   async function fetchArticles() {
     try {
-      const { data } = await storyapi.get('cdn/stories/', {
-        ...storiesConfig,
-        starts_with: 'blog/',
-        is_startpage: false,
+      const { data } = await useFetch('/api/stories/', {
+        params: {
+          starts_with: 'blog/',
+          is_startpage: false,
+        },
       })
-      state.articles = data.stories.map(formatArticleStory)
-    } catch (error) {
-      logError('There was an error fetching articles from Storyblok', error)
+      state.articles = data.value.map(formatArticleStory)
+    } catch (errorMsg) {
+      error('There was an error fetching articles from Storyblok', errorMsg)
     }
   }
 
-  async function fetchArticleBySlug(slug: string): Promise<ArticleStory> {
+  async function fetchArticleBySlug(slug: string) {
     try {
-      const { data } = await storyapi.get('cdn/stories', {
-        ...storiesConfig,
-        starts_with: 'blog/',
-        // Prepend */ to match with the first part of the full_slug
-        by_slugs: '*/' + slug,
+      const { data: story } = await useFetch(`/api/stories/blog/${slug}`, {
+        params: {
+          resolve_relations: 'category',
+        },
       })
-      const story = data.stories[0]
-      story.content.category = data.rels.find(({ uuid }) => story.content.category === uuid)
-      return formatArticleStory(story)
-    } catch (error) {
-      logError(`There was an error fetching article ${slug} from Storyblok`, error)
+
+      /* story.content.category = data.rels.find(({ uuid }) => story.content.category === uuid) */
+      return formatArticleStory(story.value)
+    } catch (errorMsg) {
+      error(`There was an error fetching article ${slug} from Storyblok`, errorMsg)
     }
   }