+++ apps/portfolio/composables/useRepositories.ts
@@ -1,9 +1,9 @@
+import { useLogger } from '@alvarosabu/use'
 import { computed } from 'vue'
 import { RemovableRef, useStorage } from '@vueuse/core'
 import { differenceInHours } from 'date-fns'
 
 import { snakeToCamel } from '@alvarosabu/utils'
-import { useLogger } from './useLogger'
 
 export enum Language {
   JavaScript = 'JavaScript',
@@ -72,7 +72,7 @@ const state: Partial<GithubState> = reactive({
   error: false,
 })
 
-const { logError } = useLogger()
+const { error } = useLogger('[ AS 🐧]')
 
 export const orderByStarsDesc = (a: GithubRepo, b: GithubRepo) => b.stars - a.stars
 
@@ -127,6 +127,11 @@ export const useGithubRepo = () => {
               per_page: 20,
             },
           }),
+          useFetch('https://api.github.com/orgs/tresjs/repos', {
+            params: {
+              per_page: 20,
+            },
+          }),
         ])
         preservedState.value.updatedAt = new Date()
         state.pending = colorsPending.value || usersPending.value || orgsPending.value
@@ -152,8 +157,8 @@ export const useGithubRepo = () => {
             })
             .sort(orderByStarsDesc)
         }
-      } catch (error) {
-        logError('There was an error fetching github repos', error)
+      } catch (errorMsg) {
+        error('There was an error fetching github repos', errorMsg)
       }
     }
   }
@@ -161,7 +166,7 @@ export const useGithubRepo = () => {
   const listRelevantRepos = computed(() =>
     preservedState.value.repositories
       .filter((repo: GithubRepo) => !repo.archived && !repo.fork && repo.stars > 1)
-      .slice(0, 3),
+      .slice(0, 6),
   )
 
   return {