+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @alvarosabu/ui
 
+## 2.19.0
+
+### Minor Changes
+
+- 025f1f8e: new Logger composable from uses package
+
 ## 2.18.1
 
 ### Patch Changes